pyproject.toml
[project]
name = "prompt-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.25.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.5.0",
    "gspread>=6.2.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
]
